tests/baseline/protocol-1.2/inspector/console/console-dir-es6-expected.txt

CONSOLE MESSAGE: line 15: [object Object]
CONSOLE MESSAGE: line 18: Symbol()
CONSOLE MESSAGE: line 22: [object Map]
CONSOLE MESSAGE: line 22: [object WeakMap]
CONSOLE MESSAGE: line 26: [object Set]
CONSOLE MESSAGE: line 26: [object WeakSet]
CONSOLE MESSAGE: line 34: [object Set]
CONSOLE MESSAGE: line 37: [object WeakMap]
CONSOLE MESSAGE: line 45: [object Map Iterator]
CONSOLE MESSAGE: line 46: [object Map Iterator]
CONSOLE MESSAGE: line 47: [object Map Iterator]
CONSOLE MESSAGE: line 45: [object Set Iterator]
CONSOLE MESSAGE: line 46: [object Set Iterator]
CONSOLE MESSAGE: line 47: [object Set Iterator]
CONSOLE MESSAGE: line 85: class{method(){ return 1 }},class        classWithWhitespace     {       method(){ return 1 }     },class FooClass {
        jump(x) { return 1 }
        badArrow(x = a => 2) { return "looooooooooooooooooooooooooooooooooooooooooooooooooooong" }
    },jump(x) { return 1 },class BarClass extends FooClass{},class BarClass2 extends class base{} {},class BarClass3 extends function base2(a, b) {} {},_ => { return 1 },(x) => { return 1 },(x, y, z) => { return 1 },({}) => { return 1 },([]) => { return 1 },() => { return "short" },() => { return "looooooooooooooooooooooooooooooooooooooooooooooooooooong" },(...x) => { return 1 },(x, y, ...z) => { return 1 },function (...x) { return 1 },function (x, y, ...z) { return 1 },function ({a}){ return 1 },function ([a]){ return 1 },function ({a, b}){ return 1 },function (...{a}){ return 1 },function (a = (1), b){ return 1 },function (a = {x: (1)}, b){ return 1 },function (a = (x) => { return 1 }, b){ return 2 },function ({a: b}){ return 1 },function (c = ")", {a: b}){ return 1 }
CONSOLE MESSAGE: line 93: badArrow(x = a => 2) { return "looooooooooooooooooooooooooooooooooooooooooooooooooooong" },function (a = ") {", b){ return 1 },function (a = function(){ return 1 }, b){ return 2 },function (a = class{ constructor(){} }){ return 2 }
Tests that console logging dumps proper messages.

console-dir-es6.html:15 Object
    a: 1
    Symbol(): 2
    Symbol(Symbol.iterator): Symbol(foo)
    Symbol(a): 3
    Symbol(a): Symbol(Symbol.iterator)
    __proto__: Object
console-dir-es6.html:18 Symbol()
console-dir-es6.html:22 Map
    size: (...)
    __proto__: Map
    [[Entries]]: Array[1]
        0: {Object => Object}
        length: 1
console-dir-es6.html:22 WeakMap
    __proto__: WeakMap
    [[Entries]]: Array[1]
        0: {Object => Object}
        length: 1
console-dir-es6.html:26 Set
    size: (...)
    __proto__: Set
    [[Entries]]: Array[1]
        0: Object
        length: 1
console-dir-es6.html:26 WeakSet
    __proto__: WeakSet
    [[Entries]]: Array[1]
        0: Object
        length: 1
console-dir-es6.html:34 Set
    size: (...)
    __proto__: Set
    [[Entries]]: Array[1]
        0: Set
        length: 1
console-dir-es6.html:37 WeakMap
    __proto__: WeakMap
    [[Entries]]: Array[0]
        length: 0
console-dir-es6.html:45 MapIterator
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "keys"
    [[Entries]]: Array[1]
        0: Object
        length: 1
console-dir-es6.html:46 MapIterator
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "values"
    [[Entries]]: Array[1]
        0: Object
        length: 1
console-dir-es6.html:47 MapIterator
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "entries"
    [[Entries]]: Array[1]
        0: Array[2]
        length: 1
console-dir-es6.html:45 SetIterator
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "values"
    [[Entries]]: Array[1]
        0: Object
        length: 1
console-dir-es6.html:46 SetIterator
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "values"
    [[Entries]]: Array[1]
        0: Object
        length: 1
console-dir-es6.html:47 SetIterator
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "entries"
    [[Entries]]: Array[1]
        0: Array[2]
        length: 1
console-dir-es6.html:85 Array[27]
    0: class 
    1: class classWithWhitespace
    2: class FooClass
    3: function jump(x)
    4: class BarClass
    5: class BarClass2
    6: class BarClass3
    7: _ => { return 1 }
    8: (x) => { return 1 }
    9: (x, y, z) => { return 1 }
    10: ({}) => { return 1 }
    11: ([]) => { return 1 }
    12: () => { return "short" }
    13: () => {…}
    14: (...x) => { return 1 }
    15: (x, y, ...z) => { return 1 }
    16: function (...x)
    17: function (x, y, ...z)
    18: function ({a})
    19: function ([a])
    20: function ({a, b})
    21: function (...{a})
    22: function (a = (1), b)
    23: function (a = {x: (1)}, b)
    24: function (a = (x) => { return 1 }, b)
    25: function ({a: b})
    26: function (c = ")", {a: b})
    length: 27
    __proto__: Array[0]
console-dir-es6.html:93 Array[4]
    0: badArrow(x = a => {…}
    1: function (a = ")
    2: function (a = function()
    3: function (a = class{ constructor()
    length: 4
    __proto__: Array[0]
